refactoring MVVM1
