+++ MVVMToy/MVVMToy/MVVM1/Model/Person.swift
@@ -1,0 +1,19 @@
+//
+//  Person.swift
+//  MVVMToy
+//
+//  Created by 정재인 on 2020/05/13.
+//  Copyright © 2020 정재인. All rights reserved.
+//
+
+import Foundation
+
+
+struct Person {
+    let firstName: String
+    let lastName: String
+    let gender: String
+    let age: Int
+    let height: Double
+    
+}
